refactor scene-manager / and ember-three-scene
